feat: connection of keyclock and created services

+++ src/main/java/com/knitwit/repository/CourseRequestRepository.java
@@ -1,9 +1,9 @@
 package com.knitwit.repository;
 
-import com.knitwit.entity.CourseRequest;
+import com.knitwit.model.CourseRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CourseRequestRepository extends JpaRepository<CourseRequest, Long> {
+public interface CourseRequestRepository extends JpaRepository<CourseRequest, Integer> {
 }
