+++ src/main/java/com/knitwit/repository/CourseSectionRepository.java
@@ -1,9 +1,10 @@
 package com.knitwit.repository;
 
-import com.knitwit.entity.CourseSection;
+import com.knitwit.model.CourseSection;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CourseSectionRepository extends JpaRepository<CourseSection, Long> {
+public interface CourseSectionRepository extends JpaRepository<CourseSection, Integer> {
+
 }
