+++ src/main/java/com/knitwit/repository/LearningProgressRepository.java
@@ -1,9 +1,9 @@
 package com.knitwit.repository;
 
-import com.knitwit.entity.LearningProgress;
+import com.knitwit.model.LearningProgress;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LearningProgressRepository extends JpaRepository<LearningProgress, Long> {
+public interface LearningProgressRepository extends JpaRepository<LearningProgress, Integer> {
 }
