+++ src/main/java/com/knitwit/repository/MediaFileRepository.java
@@ -1,9 +1,9 @@
 package com.knitwit.repository;
 
-import com.knitwit.entity.MediaFile;
+import com.knitwit.model.MediaFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MediaFileRepository extends JpaRepository<MediaFile, Long> {
+public interface MediaFileRepository extends JpaRepository<MediaFile, Integer> {
 }
